requirements.txt
Flask==2.2.2
Jinja2==3.1.2
Werkzeug==2.2.2
numpy==1.24.1
scikit-learn==0.22.1
pandas==1.5.2
tensorflow==2.10.1
keras==2.10.0
opencv-python==4.6.0.66
Pillow==9.3.0
